feat: Shortcut for response validation when Schemathesis's data generation is not used
Ref: #485
