+++ test/runner/test_checks.py
@@ -266,6 +266,7 @@ def test_response_schema_conformance_swagger(swagger_20, content, definition):
     response = make_response(content)
     case = make_case(swagger_20, definition)
     assert response_schema_conformance(response, case) is None
+    assert case.endpoint.is_response_valid(response)
 
 
 def test_response_schema_conformance_swagger_no_content_header(swagger_20):
@@ -324,6 +325,7 @@ def test_response_schema_conformance_openapi(openapi_30, content, definition):
     response = make_response(content)
     case = make_case(openapi_30, definition)
     assert response_schema_conformance(response, case) is None
+    assert case.endpoint.is_response_valid(response)
 
 
 @pytest.mark.parametrize(
@@ -338,6 +340,7 @@ def test_response_schema_conformance_invalid_swagger(swagger_20, content, defini
     case = make_case(swagger_20, definition)
     with pytest.raises(AssertionError) as exc_info:
         response_schema_conformance(response, case)
+    assert not case.endpoint.is_response_valid(response)
     assert "SchemaValidationError" in exc_info.type.__name__
 
 
@@ -367,6 +370,7 @@ def test_response_schema_conformance_invalid_openapi(openapi_30, content, defini
     case = make_case(openapi_30, definition)
     with pytest.raises(AssertionError):
         response_schema_conformance(response, case)
+    assert not case.endpoint.is_response_valid(response)
 
 
 @pytest.mark.hypothesis_nested
@@ -379,6 +383,7 @@ def test_response_schema_conformance_references_invalid(complex_schema):
         response = make_response(json.dumps({"foo": 1}).encode())
         with pytest.raises(AssertionError):
             case.validate_response(response)
+        assert not case.endpoint.is_response_valid(response)
 
     test()
